Refactor api and game routes outside of the server.js file
