+++ server.js
@@ -7,14 +7,23 @@ const io          = require('socket.io')(server);
 const request     = require('request');
 const bodyParser  = require('body-parser');
 const mongoose    = require('mongoose');
+
+var users     = [];
+let addedUser = false;
+
 const userRoutes  = require('./routes/userRoutes');
+const gameRoutes  = require('./routes/gameRoutes');
+const apiRoutes   = require('./routes/apiRoutes');
 const Game        = require('./public/js/game');
 
 // set up port that our server will be using
 app.set('port', 3000);
 
+
 app.use(bodyParser.json());
 app.use('/', userRoutes);
+app.use('/game', gameRoutes);
+app.use('/api', apiRoutes);
 
 app.use(express.static('public'));
 
@@ -26,9 +35,6 @@ mongoose.connect('mongodb://localhost/giphy', (err) => {
     console.log('Mongo connection successful');
   }
 });
-
-let users     = [];
-let addedUser = false;
 
 // ===========================================================================
 // Emit events ===============================================================
@@ -69,89 +75,6 @@ io.on('connection', (socket) => {
       });
     }
   });
-});
-
-// ===========================================================================
-// Set up routes =============================================================
-// ===========================================================================
-
-// go to homepage
-app.get('/', (req, res) => {
-  console.log('get \'/\'');
-});
-
-// go into our database and get random terms
-// the 'numberOfTerms' is how many documents from the
-// database that you want to get
-app.get('/randomTerms/:numberOfTerms', (req, res) => {
-  console.log('get /randomTerms/:numberOfTerms');
-});
-
-// hit the Giphy API and grab random giphys
-// grab all of the array of player cards,
-// and retrieve the image_url's from giphy
-// http://api.giphy.com/v1/gifs/search?q=cat&limit=1&api_key=dc6zaTOxFJmzC
-app.get('/createCards', (req, res) => {
-  console.log('get /createCards');
-
-  let searchTerm = req.query.search;
-  let searchURL = 'http://api.giphy.com/v1/gifs/search?q='
-                  + searchTerm + '&limit=1&api_key=dc6zaTOxFJmzC';
-
-  request(searchURL, (err, response, body) => {
-    let info = JSON.parse(body);
-    let giphyArray = [];
-
-    // giphyArray will hold onto the ID, GIF, and still image
-    giphyArray[0] = info['data'][0]['id'];
-    giphyArray[1] = info['data'][0]['images']['fixed_height']['url'];
-    giphyArray[2] = info['data'][0]['images']['fixed_height_still']['url'];
-
-    res.send(giphyArray);
-  });
-
-});
-
-// hit the 'Cards Against Humanity' API
-// http://www.crhallberg.com/cah/json
-// grab a list of questions and save it into our database
-app.get('/createQuestions', (req, res) => {
-  console.log('get /createQuestions');
-});
-
-// Start Round
-app.get('/startRound', (req, res) => {
-  Game.startRound(users);
-  io.emit('start round');
-});
-
-// Show user's hand
-app.get('/showHand/:userName', (req, res) => {
-  let userName = req.params.userName;
-  let myHand = Game.showHand(userName);
-
-  let searchTerm = myHand[0];
-  let searchURL = 'http://api.giphy.com/v1/gifs/search?q='
-                  + searchTerm + '&limit=1&api_key=dc6zaTOxFJmzC';
-
-  request(searchURL, (err, response, body) => {
-    let info = JSON.parse(body);
-    let giphyArray = [];
-    let handImage;
-
-    // giphyArray will hold onto the ID, GIF, and still image
-    giphyArray[0] = info['data'][0]['id'];
-    giphyArray[1] = info['data'][0]['images']['fixed_height']['url'];
-    giphyArray[2] = info['data'][0]['images']['fixed_height_still']['url'];
-
-    handImage = giphyArray[1];
-
-    console.log('Server.js show hand: ' + myHand);
-    io.emit('show hand', handImage);
-
-    res.send(giphyArray);
-  });
-
 });
 
 // set up server
